api/requirements.txt
asyncio
jetpack-io
fastapi
uvicorn
python-dotenv
redis
requests
